+++ Library/Homebrew/compat.rb
@@ -1,6 +1,7 @@
 require "compat/os/mac"
 require "compat/dependable"
 require "compat/dependency_collector"
+require "compat/download_strategy"
 require "compat/fileutils"
 require "compat/formula_support"
 require "compat/cask"
