+++ docs/Formula-Cookbook.md
@@ -546,7 +546,7 @@ class Python3 < Formula
   head "https://hg.python.org/cpython", :using => :hg
 ```
 
-Homebrew offers both anonymous and authenticated download strategies.
+Homebrew offers anonymous download strategies.
 
 | `:using` value | download strategy             |
 |----------------|-------------------------------|
@@ -561,19 +561,13 @@ Homebrew offers both anonymous and authenticated download strategies.
 | `:svn`         | `SubversionDownloadStrategy`  |
 |----------------|-------------------------------|
 
-| `:using` value            | download strategy                                | authentication source                        |
-|---------------------------|--------------------------------------------------|----------------------------------------------|
-| `:github_private_release` | `GitHubPrivateRepositoryReleaseDownloadStrategy` | `HOMEBREW_GITHUB_API_TOKEN`                  |
-| `:github_private_repo`    | `GitHubPrivateRepositoryDownloadStrategy`        | `HOMEBREW_GITHUB_API_TOKEN`                  |
-| `:s3`                     | `S3DownloadStrategy`                             | `AWS_ACCESS_KEY_ID`, `AWS_SECRET_ACCESS_KEY` |
-| `:scp`                    | `ScpDownloadStrategy`                            | SSH key pair                                 |
-|---------------------------|--------------------------------------------------|----------------------------------------------|
-
 If you need more control over the way files are downloaded and staged, you can create a custom download strategy and specify it using the [`url`](https://www.rubydoc.info/github/Homebrew/brew/master/Formula#url-class_method) method's `:using` option:
 
 ```ruby
 class MyDownloadStrategy < SomeHomebrewDownloadStrategy
-  # Does something cool
+  def fetch
+    # downloads output to `temporary_path`
+  end
 end
 
 class Foo < Formula
